# openapi/spec.yaml
openapi: "3.0.3"
info:
  title: ADH6 API
  description: >
    This is the specification for **MiNET**'s ADH6 plaform. Its aim is to manage our users, devices and treasury.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
  contact:
    name: MiNET
    email: equipe@minet.net
    url: https://minet.net
servers:
  - url: https://adh6.minet.net/api
    description: ADH6 production server
  - url: https://adh6-dev.minet.net/api
    description: ADH6 staging server
  - url: https://adh6-local.minet.net/api
    description: ADH6 staging server
paths:
...
  /device/{id}/member:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
                example: 0
                nullable: true
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Get user of the device
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - device
  /member:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
        - description: Filters by various properties
          in: query
          name: filter
          schema:
            $ref: "#/components/schemas/MemberFilter"
          style: deepObject
          explode: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: integer
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Filter members
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - member
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberBody"
        description: The member to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                description: The id of the created user
                type: integer
                example: 1
          description: The created account
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Create a member
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - member
  /member/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    delete:
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Delete a member
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - member
    get:
      parameters:
        - $ref: "#/components/parameters/onlyMember"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbstractMember"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve a member
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - member
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberBody"
        description: The new values for this member
        required: true
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Partially update a member
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
      - member
  /member/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      parameters:
        - description: Whether to fetch DHCP logs
          in: query
          name: dhcp
          required: false
          schema:
            type: boolean
        - description: Maximum number of logs to retrieve
          in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - description: Number of logs to skip (for pagination)
          in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                          format: date-time
                          example: "2018-07-14T04:44:51.537Z"
                        message:
                          type: string
                          example: "Password OK"
                      required:
                        - timestamp
                        - message
                  total:
                    type: integer
                    description: Total number of logs available
                    example: 150
                  hasMore:
                    type: boolean
                    description: Whether there are more logs available
                    example: true
                required:
                  - logs
                  - total
                  - hasMore
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: '#/components/responses/Unexpected'
      summary: Retrieve the most recent logs of a member
      security: 
        - OpenIdConnect:
          - admin:read
      tags:
      - member
  /member/{id}/statuses:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MemberStatus"
                type: array
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieves some common status updates concerning a member
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - member
  /member/{id}/password:
    parameters:
      - $ref: "#/components/parameters/idPath"
    put:
      requestBody:
        content:
          application/json:
            schema:
              properties:
                password:
                  example: "&<Hs8XK/H+dT{7d7:=bxN~9;EqhwtW-;"
                  type: string
                  format: password
                  description: The plaintext password to use
                hashedPassword:
                  example: "de6a3b882c1cff12af8616e1f6bf5b6c"
                  type: string
                  format: md4
                  description: The md4-hashed password to use. MD4 is obv. long-deprecated but we use NTLM for PEAP authentication... WARNING. Hashing the password client-side essentially nullifies the purpose of hashing the password.
                  deprecated: true
              type: object
        description: The new value for the password in plaintext
        required: true
      responses:
        "204":
          description: Modified
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Update the password of a member
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth:
            - admin:write
      tags:
        - member
  /member/{id}/charter/{charter_id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - description: "The unique identifier of the charter: 1 for MiNET and 2 for Hosting"
        in: path
        name: charter_id
        required: true
        schema:
          type: integer
          enum:
            - 1
            - 2
    get:
      deprecated: true
      operationId: charter_get
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
                format: date-time
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieves if the hosting charter has been signed
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - member
    put:
      deprecated: true
      operationId: charter_put
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Update if the hosting/MiNET charter has been signed
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
      - member
  /member/{id}/comment:
    parameters:
      - $ref: "#/components/parameters/idPath"
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        "204":
          description: Modified
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Update the comment of a member
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth:
            - admin:write
      tags:
        - member
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieves the comment of the user
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth:
            - admin:read
      tags:
      - member
...
  /member/{id}/subscription/validate:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - description: Should the membership be free
        in: query
        name: free
        schema:
          type: boolean
          example: false
    post:
      operationId: subscription_validate
      responses:
        "204":
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Validate a pending transaction
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
      - membership
...
components:
  parameters:
...
    idPath:
      description: The id of the account that needs to be fetched.
      in: path
      name: id
      required: true
      schema:
        type: integer
        example: 0
...
  securitySchemes:
    OpenIdConnect:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/auth
          tokenUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/token
          scopes:
            user: Basic user access
            admin:read: Read access to admin resources
            admin:write: Write access to admin resources
            admin:prod: Production admin access
            treasurer:read: Read access to treasury resources
            treasurer:write: Write access to treasury resources
            network:read: Read access to network resources
            network:write: Write access to network resources
            network:write:prod: Production network write access
            network:write:dev: Development network write access
            network:write:hosting: Hosting network write access
    ApiKeyAdminAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_admin_auth
    ApiKeyNetworkAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_network_auth
    ApiKeyTreasuryAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_treasury_auth
...